refactor(repository): migrated countries to its own folder
